initial annotation review functionality

+++ modules/LabelUI/backend/middleware.py
@@ -96,7 +96,10 @@ class DBMiddleware():
                 response[imgID]['viewcount'] = viewcount
             last_checked = b['last_checked']
             if last_checked is not None:
-                response[imgID]['last_checked'] = last_checked.timestamp()
+                if response[imgID]['last_checked'] is None:
+                    response[imgID]['last_checked'] = last_checked
+                else:
+                    response[imgID]['last_checked'] = max(response[imgID]['last_checked'], last_checked)
 
             # parse annotations and predictions
             entryID = str(b['id'])
@@ -115,7 +118,7 @@ class DBMiddleware():
                     response[imgID]['annotations'][entryID] = entry
                 elif b['ctype'] == 'prediction':
                     response[imgID]['predictions'][entryID] = entry
-        
+
         return response
 
 
@@ -249,45 +252,6 @@ class DBMiddleware():
         return { 'entries': response }
         
 
-    def getBatch_timeRange(self, username, minTimestamp, maxTimestamp, userList, limit=None):
-        '''
-            Returns images that have been annotated within the given time range and/or
-            by the given user(s). All arguments are optional.
-            Useful for reviewing existing annotations.
-        '''
-        # query string
-        sql = self.sqlBuilder.getDateQueryString(minTimestamp, maxTimestamp, userList)
-
-        # check validity and provide arguments
-        queryVals = []
-        if userList is not None:
-            queryVals.append(userList)
-        if minTimestamp is not None:
-            queryVals.append(minTimestamp)
-        if maxTimestamp is not None:
-            queryVals.append(maxTimestamp)
-
-        # limit (TODO: make 128 a hyperparameter)
-        if limit is None:
-            limit = 128
-        else:
-            limit = min(int(limit), 128)
-        queryVals.append(limit)
-
-        # query and parse results
-        with self.dbConnector.execute_cursor(sql, tuple(queryVals)) as cursor:
-            try:
-                response = self._assemble_annotations(cursor)
-                # self.dbConnector.conn.commit()
-            except Exception as e:
-                print(e)
-                # self.dbConnector.conn.rollback()
-            finally:
-                pass
-                # cursor.close()
-        return { 'entries': response }
-
-    
     def getBatch_auto(self, username, order='unlabeled', subset='default', limit=None):
         '''
             TODO: description
@@ -310,6 +274,73 @@ class DBMiddleware():
             response = self._assemble_annotations(cursor)
 
         return { 'entries': response }
+
+
+    def getBatch_timeRange(self, minTimestamp, maxTimestamp, userList, skipEmptyImages=False, limit=None):
+        '''
+            Returns images that have been annotated within the given time range and/or
+            by the given user(s). All arguments are optional.
+            Useful for reviewing existing annotations.
+        '''
+        # query string
+        sql = self.sqlBuilder.getDateQueryString(minTimestamp, maxTimestamp, userList, skipEmptyImages)
+
+        # check validity and provide arguments
+        queryVals = []
+        if userList is not None:
+            queryVals.append(tuple(userList))
+        if minTimestamp is not None:
+            queryVals.append(minTimestamp)
+        if maxTimestamp is not None:
+            queryVals.append(maxTimestamp)
+        if skipEmptyImages and userList is not None:
+            queryVals.append(tuple(userList))
+
+        # limit (TODO: make 128 a hyperparameter)
+        if limit is None:
+            limit = 128
+        else:
+            limit = min(int(limit), 128)
+        queryVals.append(limit)
+
+        if userList is not None:
+            queryVals.append(tuple(userList))
+
+        # query and parse results
+        with self.dbConnector.execute_cursor(sql, tuple(queryVals)) as cursor:
+            try:
+                response = self._assemble_annotations(cursor)
+                # self.dbConnector.conn.commit()
+            except Exception as e:
+                print(e)
+                # self.dbConnector.conn.rollback()
+            finally:
+                pass
+                # cursor.close()
+        return { 'entries': response }
+
+    
+    def get_timeRange(self, userList, skipEmptyImages=False):
+        '''
+            Returns two timestamps denoting the temporal limits within which
+            images have been viewed by the users provided in the userList.
+            Arguments:
+            - userList: string (single user name) or list of strings (multiple).
+                        Can also be None; in this case all annotations will be
+                        checked.
+            - skipEmptyImages: if True, only images that contain at least one
+                               annotation will be considered.
+        '''
+        # query string
+        sql = self.sqlBuilder.getTimeRangeQueryString(userList, skipEmptyImages)
+
+        arguments = (None if userList is None else (userList,))
+        result = self.dbConnector.execute(sql, arguments, numReturn=1)
+
+        return {
+            'minTimestamp': result[0]['mintimestamp'],
+            'maxTimestamp': result[0]['maxtimestamp'],
+        }
 
 
     def submitAnnotations(self, username, submissions):
@@ -393,23 +424,24 @@ class DBMiddleware():
 
         # delete all annotations that are not in submitted batch
         imageKeys = list(UUID(k) for k in submissions['entries'])
-        if len(ids):
-            sql = '''
-                DELETE FROM {schema}.annotation WHERE username = %s AND id IN (
-                    SELECT idQuery.id FROM (
-                        SELECT * FROM {schema}.annotation WHERE id NOT IN %s
-                    ) AS idQuery
-                    JOIN (
-                        SELECT * FROM {schema}.annotation WHERE image IN %s
-                    ) AS imageQuery ON idQuery.id = imageQuery.id);
-            '''.format(schema=schema)
-            self.dbConnector.execute(sql, (username, tuple(ids), tuple(imageKeys),))
-        else:
-            # no annotations submitted; delete all annotations submitted before
-            sql = '''
-                DELETE FROM {schema}.annotation WHERE username = %s AND image IN %s;
-            '''.format(schema=schema)
-            self.dbConnector.execute(sql, (username, tuple(imageKeys),))
+        if len(imageKeys):
+            if len(ids):
+                sql = '''
+                    DELETE FROM {schema}.annotation WHERE username = %s AND id IN (
+                        SELECT idQuery.id FROM (
+                            SELECT * FROM {schema}.annotation WHERE id NOT IN %s
+                        ) AS idQuery
+                        JOIN (
+                            SELECT * FROM {schema}.annotation WHERE image IN %s
+                        ) AS imageQuery ON idQuery.id = imageQuery.id);
+                '''.format(schema=schema)
+                self.dbConnector.execute(sql, (username, tuple(ids), tuple(imageKeys),))
+            else:
+                # no annotations submitted; delete all annotations submitted before
+                sql = '''
+                    DELETE FROM {schema}.annotation WHERE username = %s AND image IN %s;
+                '''.format(schema=schema)
+                self.dbConnector.execute(sql, (username, tuple(imageKeys),))
 
         # insert new annotations
         if len(values_insert):
